Add additional repositories, exceptions, DTO's

+++ pharmacy/src/main/java/org/pharmacy/repository/ClientRepository.java
@@ -1,6 +1,9 @@
 package org.pharmacy.repository;
 
+import org.pharmacy.exceptions.DataNotFoundException;
+import org.pharmacy.model.Address;
 import org.pharmacy.model.Client;
+import org.pharmacy.exceptions.DataIntegrityViolationException;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -17,52 +20,60 @@ public class ClientRepository {
     }
 
     public void addClient(Client client) throws SQLException {
-        if (client.getFirstname() == null || client.getFirstname().trim().isEmpty()
-                || client.getLastname() == null || client.getLastname().trim().isEmpty()
-                || client.getCountry() == null || client.getCountry().trim().isEmpty()
-                || client.getCity() == null || client.getCity().trim().isEmpty()
-                || client.getStreet() == null || client.getStreet().trim().isEmpty()
-                || client.getPostalCode() == null || client.getPostalCode().trim().isEmpty()) {
-            throw new IllegalArgumentException("All client fields must be filled.");
+        if (client == null){
+            throw new IllegalArgumentException("Client object cannot be null.");
         }
 
         final String SQLQuery = "INSERT INTO client(first_name, last_name, country, city, street, postal_code) " +
                 "VALUES (?, ?, ?, ?, ?, ?)";
 
         try (PreparedStatement pstmt = conn.prepareStatement(SQLQuery)) {
-            pstmt.setString(1, client.getFirstname());
-            pstmt.setString(2, client.getLastname());
-            pstmt.setString(3, client.getCountry());
-            pstmt.setString(4, client.getCity());
-            pstmt.setString(5, client.getStreet());
-            pstmt.setString(6, client.getPostalCode());
+            pstmt.setString(1, client.firstname());
+            pstmt.setString(2, client.lastname());
+            pstmt.setString(3, client.address().country());
+            pstmt.setString(4, client.address().city());
+            pstmt.setString(5, client.address().city());
+            pstmt.setString(6, client.address().postalCode());
 
             int affectedRows = pstmt.executeUpdate();
-            System.out.println("Client " + client.getFirstname() + " " + client.getLastname() + " successfully added. Rows changed: " + affectedRows);
+            System.out.println("Client " + client.firstname() + " " + client.lastname() + " successfully added. Rows changed: " + affectedRows);
         }
     }
 
-    public void deleteClient(Client client) throws SQLException {
+    public void deleteClient(long clientId) throws SQLException {
+        // Validation check for the ID itself (Repository responsibility)
+        if (clientId <= 0) {
+            throw new IllegalArgumentException("Client ID must be positive for deletion.");
+        }
+
         final String SQLQuery = "DELETE FROM client WHERE client_id = ?";
+
         try(PreparedStatement pstmt = conn.prepareStatement(SQLQuery)){
-            pstmt.setLong(1, client.getId());
+            pstmt.setLong(1, clientId);
             int affectedRows = pstmt.executeUpdate();
 
             if (affectedRows == 0) {
-                throw new SQLException("Client with ID " + client.getId() + " was not found.");
+                throw new DataNotFoundException("Client with ID " + clientId + " was not found.");
             }
-            System.out.println("Client with ID " + client.getId() + " successfully deleted");
+            System.out.printf("Client with ID %d successfully deleted.\n", clientId);
+
+        } catch (SQLException e) {
+            if ("23503".equals(e.getSQLState())) {
+                throw new DataIntegrityViolationException(
+                        String.format("Client with ID %d cannot be deleted because there are existing orders associated with this client. Please delete the orders first.", clientId), e);
+            }
+            throw e;
         }
     }
 
-    public void updateClientAddress(long clientId, String country, String city, String street, String postal_code) throws SQLException {
+    public void updateClientAddress(long clientId, Address address) throws SQLException {
         final String SQLQuery = "UPDATE client SET country = ?, city = ?, street = ?, postal_code = ? WHERE client_id = ?";
 
         try(PreparedStatement pstmt = conn.prepareStatement(SQLQuery)){
-            pstmt.setString(1, country);
-            pstmt.setString(2, city);
-            pstmt.setString(3, street);
-            pstmt.setString(4, postal_code);
+            pstmt.setString(1, address.country());
+            pstmt.setString(2, address.city());
+            pstmt.setString(3, address.street());
+            pstmt.setString(4, address.postalCode());
             pstmt.setLong(5, clientId);
 
             int affectedRows = pstmt.executeUpdate();
@@ -77,15 +88,19 @@ public class ClientRepository {
         try (PreparedStatement pstmt = conn.prepareStatement(SQLQuery);
              ResultSet rs = pstmt.executeQuery()) {
             while (rs.next()) {
+                Address address = new Address(
+                        rs.getString("country"),
+                        rs.getString("city"),
+                        rs.getString("street"),
+                        rs.getString("postal_code")
+                );
+
                 clients.add(new Client(
                         rs.getLong("client_id"),
                         rs.getString("first_name"),
                         rs.getString("last_name"),
-                        rs.getString("country"),
-                        rs.getString("city"),
-                        rs.getString("street"),
-                        rs.getString("postal_code"))
-                );
+                        address
+                ));
             }
         }
         return clients;
